refactor(1013): remove OrgDeviceInventorySummary facade (Cat B, position 29)
Extract the OrgDeviceInventorySummary delegation facade from MistHelper.py
into src/inventory/org_device_inventory_summary_facade.py. The class was
already a thin dispatch layer over the extracted
src.inventory.org_device_inventory_summary (single-org core) and
src.inventory.org_device_inventory_msp (MSP orchestrator) modules;
this change moves the facade itself out of the monolith.

Live-global reads (apisession, mistapi, DataExporter, org_id, InputUtils,
msp_privileges) now resolve via lazy `mh = importlib.import_module("MistHelper")`
inside each helper, matching the SC-001 pattern used across the Cat B block.

Callers keep the `MistHelper.OrgDeviceInventorySummary` API through the
alphabetically-inserted re-export alias in the src.<domain>.* import block.
Coverage-omit entry added for the new module.

Local gate: black + ruff clean; py_compile OK; --test baseline 66/0/131.
